Fixed the handling of the return message from IO HTTP bind
We were not sending this message from within an actor, and previously nobody was handling the response. This was resulting in dead letters in the case of success, and provided no handling in the case of failure.

Blocking on the result of my ask to avoid potential for down the road concurrency issues.