Added an intial implementation of a CLI program.
Currently handles only methods push. Ultimately will want fetch as well. Also support for task-configurations.

Method push mimics git commit. Brings up a text editor to get user input for the synopsis/documentation metadata.

We are currently assuming you have gotten the auth token as a string, and are supplying this on command line (this is what was done for apollo-testing). To make this really usable we need to enable programatic login to the OpenAm server to grab the token.

I'm not happy with the arg parsing/validation here. Grabs some defaults from a .apolloconfig, assumed to be in the same directory. Also meant to sort of mimic git. This is mostly for convenience, to save specifying from specifying user + namespace every single time. Also to set the target url (for CI vs Prod vs Dev deployments). But the validation here is a little ugly. Also, I'd like to enable multiple 'modes', push vs fetch, that carry a different set of required arguments. Not sure how to do this, but I'd love input.