Working version of method read authorization restriction
GcsAuthorizaitonProvider is making spray-client calls to GCS JSON API endpoints. 404 responses for ACL get are interpreted as Nothing permission.

For the method/config list endpoint, we collect the http response futures into a single future of a sequence, and await the result of that.
This allows futures to complete in parallel, ensuring we only block until the slowest http response arrives.

Bubbling GCS contact failures up to the query-handler as a ClientServiceFailure exception, which the queryHandler catches. Responds to user as 500 error.

Moved a ton of stuff into new packages

This version works on pre-created buckets.
Not sure this is the best way to handle the refresh token when authenticating GCS contact